backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
face-recognition==1.3.0
numpy==1.26.4
Pillow==10.4.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
